requirements.txt
# Hierarchical Todo App - Python Dependencies
# ============================================
# 
# Backend dependencies for the Flask API server.
# These are the core packages needed to run the application.
# 
# Installation: pip install -r requirements.txt

# Flask web framework and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1  # Database ORM
Flask-Login==0.6.3       # User session management
Flask-Cors==4.0.0        # Cross-origin resource sharing

# Environment and configuration
python-dotenv==1.0.0     # Environment variable loading